service1/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
pyjwt==2.8.0
bcrypt==4.0.1
requests==2.28.1